Updated GA in search.ipynb
Removed the image links and included the images used in the image
folder. Reduced the file size from 2.4 MB to 183 KB. For every 21 print
statements that we previously had, now we have 2, only printing the
fittest individual in each generation. The instructions has been made to
get the detailed output.