Add test for Table-Driven-Vacuum-Agent.
Note that this changes the percept sequence table
in the code, it now slightly differs from table 2.3
in the book. The underlying problem here is that a
full table can principally not be specified, so '...'
is used to represent the missing table entries. While
this works for the book, the code requires some
additional table entries for the agent to succeed.