Updated display.py:
Implemented an Icon object to handle the details of moving compound images around a Tkinter canvas.  Removed previous implementation of images from the code.

Updated left_click() method to display the performance of any agents in the cell when clicked.

Created right_click() method to display the held items of any agents in the cell when clicked.

Created an object.destroy() method to hold class-specific destruction code.  This is extended when used in a display object to include Icon and image removal code.