Updated vacuum.py:
Added a agent.performance bonus of +100 when picking up dirt and -1 for each timestep.

Fixed an error with holding objects appending as a list not as individual items.

Removed the add_agent() method and merged the behavior in the add_object() method using if isinstance(obj, Agent):

Restructured the exogenous_change() to be configured as part of the environment definition, rather than built-in behavior as part of the class.

Configured the test_agent() code to evaluate the running of Agents without a GUI.  An example can be run under test2()

Added a seed to provide repeatable outcomes, when desired.  This is configured in the main() function.